Adjusted logic so that the Tutor and Student entitty would include a User, shifted logic of video and language lists away from the user
