+++ src/main/java/com/badar/llp/Models/Student.java
@@ -2,6 +2,7 @@ package com.badar.llp.Models;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.OneToMany;
+import jakarta.persistence.OneToOne;
 
 import java.util.List;
 
@@ -11,11 +12,32 @@ public class Student extends BasicModel {
     @OneToMany
     private List<Language> languageList;
 
+    private List<Video> videoList;
+
+    @OneToOne
+    private User user;
+
     public List<Language> getLanguageList() {
         return languageList;
     }
 
     public void setLanguageList(List<Language> languageList) {
         this.languageList = languageList;
+    }
+
+    public List<Video> getVideoList() {
+        return videoList;
+    }
+
+    public void setVideoList(List<Video> videoList) {
+        this.videoList = videoList;
+    }
+
+    public User getUser() {
+        return user;
+    }
+
+    public void setUser(User user) {
+        this.user = user;
     }
 }
